backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
websockets==13.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# AI Providers
openai==1.54.4
anthropic==0.39.0
google-generativeai==0.8.3
httpx==0.27.2

# MCP SDK
mcp==1.1.2

# Document Processing
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5

# Visualization
matplotlib==3.9.2

# Utilities
python-dotenv==1.0.1
pydantic==2.10.2
pydantic-settings==2.6.1
